Localization: move generated files
