+++ packages/devextreme/.gitignore
@@ -5,8 +5,8 @@ node_modules
 /testing/LastSuiteTime.txt
 /dotnet_packages
 /js/bundles/dx.custom.js
-/js/common/core/localization/default_messages.js
-/js/common/core/localization/cldr-data
+/js/__internal/core/localization/default_messages.ts
+/js/__internal/core/localization/cldr-data
 /scss/bundles/*.scss
 RawLog.txt
 testing/CompletedSuites.txt
