+++ packages/devextreme/js/__internal/core/localization/intl/number.ts
@@ -1,5 +1,5 @@
-import accountingFormats from '@js/common/core/localization/cldr-data/accounting_formats';
 import dxConfig from '@js/core/config';
+import accountingFormats from '@ts/core/localization/cldr-data/accounting_formats';
 import localizationCoreUtils from '@ts/core/localization/core';
 import type { FormatConfig as BaseFormatConfig, LocalizationFormat } from '@ts/core/localization/number';
 import openXmlCurrencyFormat from '@ts/core/localization/open_xml_currency_format';
