+++ packages/devextreme/js/common/core/localization/cldr-data/accounting_formats.js
@@ -1,0 +1,4 @@
+// deprecated
+import accountingFormats from '../../../../__internal/core/localization/cldr-data/accounting_formats';
+
+export default accountingFormats;
