+++ packages/devextreme/js/common/core/localization/cldr-data/en.js
@@ -1,0 +1,2 @@
+// deprecated
+export { enCldr } from '../../../../__internal/core/localization/cldr-data/en';
