+++ packages/devextreme/js/common/core/localization/cldr-data/first_day_of_week_data.js
@@ -1,0 +1,4 @@
+// deprecated
+import firstDayOfWeekData from '../../../../__internal/core/localization/cldr-data/first_day_of_week_data';
+
+export default firstDayOfWeekData;
