+++ packages/devextreme/js/common/core/localization/cldr-data/parent_locales.js
@@ -1,0 +1,4 @@
+// deprecated
+import parentLocales from '../../../../__internal/core/localization/cldr-data/parent_locales';
+
+export default parentLocales;
