+++ packages/devextreme/js/common/core/localization/cldr-data/supplemental.js
@@ -1,0 +1,2 @@
+// deprecated
+export { supplementalCldr } from '../../../../__internal/core/localization/cldr-data/supplemental';
