+++ packages/devextreme/js/common/core/localization/default_messages.js
@@ -1,0 +1,2 @@
+// deprecated
+export { defaultMessages } from '../../../__internal/core/localization/default_messages';
