+++ packages/devextreme/js/localization/cldr-data/first_day_of_week_data.js
@@ -1,1 +1,1 @@
-export { default } from '../../common/core/localization/cldr-data/first_day_of_week_data';
+export { default } from '../../__internal/core/localization/cldr-data/first_day_of_week_data';
