+++ packages/devextreme/js/localization/default_messages.js
@@ -1,1 +1,1 @@
-export * from '../common/core/localization/default_messages';
+export * from '../__internal/core/localization/default_messages';
